게시물 컴포넌트 및 컨트롤러 수정

+++ Server/controller/post.js
@@ -139,7 +139,6 @@ module.exports = {
     try {
       const postId = req.params.id;
       const { title, content } = req.body;
-
       // 게시물 작성자인지 판별한다
       const userSession = req.session.user ? req.session.user : null;
       let userId = null;
@@ -161,21 +160,36 @@ module.exports = {
       if (!isAuthor) {
         return res.status(403).send("권한이 없습니다.");
       }
+      if (req.files && req.files.length > 0) {
+        {
+          // 이전파일 삭제
+          const existingFiles = await models.db.images.findAll({
+            where: { post_id: existingPost.id },
+            transaction: t,
+          });
 
-      // 이전파일 삭제
-      const existingFiles = await models.db.images.findAll({
-        where: { post_id: existingPost.id },
-        transaction: t,
-      });
+          const fileDeletionPromises = existingFiles.map(async (file) => {
+            await models.db.images.destroy({
+              where: { id: file.id },
+              transaction: t,
+            });
+          });
+          await Promise.all(fileDeletionPromises);
 
-      const fileDeletionPromises = existingFiles.map(async (file) => {
-        await fs.unlink(file.url);
-        await models.db.images.destroy({
-          where: { id: file.id },
-          transaction: t,
-        });
-      });
-      await Promise.all(fileDeletionPromises);
+          // 첨부 파일 수정
+
+          // 새 첨부파일 업로드
+          const filePromises = req.files.map(async (file) => {
+            const fileData = {
+              url: file.location,
+              post_id: existingPost.id,
+            };
+            await models.db.images.create(fileData, { transaction: t });
+          });
+
+          await Promise.all(filePromises);
+        }
+      }
 
       // 게시물 업데이트
       await existingPost.update(
@@ -185,20 +199,6 @@ module.exports = {
         },
         { transaction: t } // 트랜잭션 지정
       );
-
-      // 첨부 파일 수정
-      if (req.files && req.files.length > 0) {
-        // 새 첨부파일 업로드
-        const filePromises = req.files.map(async (file) => {
-          const fileData = {
-            url: file.path + "/" + file.originalname,
-            post_id: existingPost.id,
-          };
-          await models.db.images.create(fileData, { transaction: t });
-        });
-
-        await Promise.all(filePromises); // 병렬 처리
-      }
 
       await t.commit();
